Add files via upload
This summer, I was a research student under Professor Almut Burchard in the Pure Mathematics Department at the University of Toronto. We worked on an open problem in geometric analysis, in the field of Steiner Symmetrizations. The problem was: does there exist an algorithm to symmetrize any set exponentially fast?

To investigate various algorithms, I wrote a Python code that would symmetrize any polygon by any angle. The user must simply provide the vertices of the polygon and the angle. The program will return the symmetrized polygon! 

This code was used to gather many (over 100) data points about various symmetrization algorithms. The data points were analyzed to determine which algorithms were fastest.